requirements.txt
selenium
python-dotenv
discord.py